+++ src/sensors/AtlasScientificCO2.h
@@ -279,8 +279,7 @@ class AtlasScientificCO2_CO2 : public Variable {
     explicit AtlasScientificCO2_CO2(
         AtlasScientificCO2* parentSense, const char* uuid = "",
         const char* varCode = ATLAS_CO2_DEFAULT_CODE)
-        : Variable(parentSense, static_cast<uint8_t>(ATLAS_CO2_VAR_NUM),
-                   static_cast<uint8_t>(ATLAS_CO2_RESOLUTION),
+        : Variable(parentSense, ATLAS_CO2_VAR_NUM, ATLAS_CO2_RESOLUTION,
                    ATLAS_CO2_VAR_NAME, ATLAS_CO2_UNIT_NAME, varCode, uuid) {}
     /**
      * @brief Destroy the AtlasScientificCO2_CO2 object - no action needed.
@@ -312,8 +311,7 @@ class AtlasScientificCO2_Temp : public Variable {
     explicit AtlasScientificCO2_Temp(
         AtlasScientificCO2* parentSense, const char* uuid = "",
         const char* varCode = ATLAS_CO2TEMP_DEFAULT_CODE)
-        : Variable(parentSense, static_cast<uint8_t>(ATLAS_CO2TEMP_VAR_NUM),
-                   static_cast<uint8_t>(ATLAS_CO2TEMP_RESOLUTION),
+        : Variable(parentSense, ATLAS_CO2TEMP_VAR_NUM, ATLAS_CO2TEMP_RESOLUTION,
                    ATLAS_CO2TEMP_VAR_NAME, ATLAS_CO2TEMP_UNIT_NAME, varCode,
                    uuid) {}
     /**
